+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/damage/Smite.java
@@ -18,9 +18,4 @@ public class Smite extends BonusDamageEnchantment {
     public boolean isAffectedByEnchantment(Entity target) {
         return target.getType().is(AFFECTED_BY_SMITE);
     }
-
-    @Override
-    public float getDamageBonusPerLevel() {
-        return 1.25f;
-    }
 }